move eos_2.py to eos.py
